`Removed unused icons and constants, updated import paths and file names`
