+++ src/pages/company/children/OpeningIcons.tsx
